backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
faiss-cpu==1.7.4
pymupdf==1.23.7
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
requests==2.31.0
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2
torch==2.1.1
groq==0.4.1
streamlit==1.28.0
tqdm==4.66.1